requirements.txt
Django==2.2.9
django-environ==0.4.5
django-extensions==2.2.6
django-simple-history==2.12.0
django-summernote==0.8.11.6
django-taggit==1.3.0
Pillow==7.1.2
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
